EF models and dependency injection

+++ src/Core/Domain/Helpers/Money.cs
@@ -1,4 +1,4 @@
 ﻿namespace Domain.Helpers
 {
-    public record Money(int Value, string Currency);
+    public record Money(double Value, string Currency);
 }
